Updates for Accumulo 2.1.0 SLF4J2 and new MIME type (#12)
Accumulo 2.1.0 recently upgraded slf4j to slf4j2. I updated the poms
to account for the new version and additional dependency. Two of the
tests were also failing, which turned out to be caused by the backport
of JDK-8273655. Adding a mime type mapping for application/java-archive
resolved the issue. While debugging I restructured the VFSClassLoaderTest.