Make test scripts clean up duplicate classes
The test scripts in start create multiple classes with the same name
that can cause problems for IDEs. The classes are only used to create
jars for the VFS tests so just drop the intermediate class files when done
creating the jars.

Note: This is the same fix used in Accumulo:
https://github.com/apache/accumulo/commit/f5121ec