Add tools for ContextDefinition create and cleanup
Modified ContextDefinition to implement KeywordExecutable so that
the user can use `accumulo create-context-definition` to create the
json required for the ContextDefinition file.

Added a Cleaner to call `URLClassLoader.close` on instances that are
about to be garbage collected. The LocalCachingContextCleaner is used
to register the URLClassLoader with the Cleaner and this class also
keeps references to the URLClassLoaders. The list is accessed from
a MXBean so that a user can use JMX to retrieve a list of files that
are referenced in the local cache directory from the classloaders
that have been created but not yet gc'd. This should aid users in
determining which files can be removed from the local cache directory.